feat(README.md): Added link to a digital editions course
Added link to a digital editions course that covers the whole process of creating a digital edition, from selecting a text right through to publication. The training sessions are interactive, the XML sessions have exercises to complete, and many of the slides have clickable hyperlinks, sometimes embedded in images. Each session has a worksheet to help users with that stage of the course. (#34)